+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [0.1.4] - 2026-04-08
+
+### Added
+- AZ023 lint rule: "Rule entry is not a dict" (ERROR)
+- AZ024 lint rule: "Phase value is not a list" (ERROR)
+
+### Changed
+- README heading capitalization normalized to sentence case
+- README em dashes standardized
+
 ## [0.1.3] - 2026-04-07
 
 ### Added